refactor bae test
